pyproject.toml
[project]
name = "rt-weather-app"
version = "1.0.0"
description = "A Flask-based weather application"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "flask>=3.0.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.0.0"
]

[project.urls]
Homepage = "https://github.com/rickytang666/rt-weather-app"
Repository = "https://github.com/rickytang666/rt-weather-app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]
